requirements.txt
boto3==1.42.87
botocore==1.42.87
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
ImageIO==2.37.3
jmespath==1.1.0
Naked==0.1.32
numpy==2.4.4
opencv-python==4.13.0.92
pillow==12.2.0
pycparser==3.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
tabulate==0.10.0
termcolor==3.3.0
urllib3==2.6.3
